Keep previous bank account balance when it changes
When we update the bank account balance, we keep the balane history by
creating of a new balance record instead of updating the value in the
current one.

Through this we manage to keep a history of account balances.

The bank account current balance will be the latest associated balance
record.